+++ integrations/cli/index.test.ts
@@ -52,9 +52,14 @@ describe.each([
             addVariant('hocus', ['&:focus', '&:hover'])
           }
         `,
+        'project-a/tailwind.config.js': js`
+          module.exports = {
+            content: ['../project-b/src/**/*.js'],
+          }
+        `,
         'project-a/src/index.css': css`
           @import 'tailwindcss/utilities';
-          @source '../../project-b/src/**/*.js';
+          @config '../tailwind.config.js';
           @plugin '../plugin.js';
         `,
         'project-a/src/index.js': js`
@@ -111,9 +116,14 @@ describe.each([
             addVariant('hocus', ['&:focus', '&:hover'])
           }
         `,
+        'project-a/tailwind.config.js': js`
+          module.exports = {
+            content: ['../project-b/src/**/*.js'],
+          }
+        `,
         'project-a/src/index.css': css`
           @import 'tailwindcss/utilities';
-          @source '../../project-b/src/**/*.js';
+          @config '../tailwind.config.js';
           @plugin '../plugin.js';
         `,
         'project-a/src/index.js': js`