+++ packages/@tailwindcss-vite/src/index.ts
@@ -93,8 +93,9 @@ export default function tailwindcss(): Plugin[] {
     })
 
     scanner = new Scanner({
-      sources: globs.map((pattern) => ({
-        base: inputBasePath, // Globs are relative to the input.css file
+      sources: globs.map(({ base, pattern }) => ({
+        // Globs are relative to the input.css file
+        base: base ? path.dirname(path.resolve(inputBasePath, base)) : inputBasePath,
         pattern,
       })),
     })