+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet!
+### Fixed
+
+- Ensure content globs defined in `@config` files are relative to the config file
 
 ## [4.0.0-alpha.21] - 2024-09-02